feat: implement Global Search Quick Apps and Results components
- Added GlobalSearchQuickAppsBar and GlobalSearchQuickAppManager for managing quick app icons.
- Introduced GlobalSearchResults component to display search results with appropriate icons and formatting.
- Updated GlobalSearchPanel tests to accommodate new search types and filters for topics and sessions.
- Refactored globalSearchGroups to separate topic and session filters, enhancing search functionality.
- Updated localization files to include new search type labels and message search entries.
