+++ src/renderer/src/i18n/locales/en-us.json
@@ -2193,7 +2193,9 @@
       "assistant": "Assistant",
       "conversation": "Conversation",
       "knowledge": "Knowledge",
-      "label": "Search type"
+      "label": "Search type",
+      "session": "Session",
+      "topic": "Topic"
     },
     "groups": {
       "agent": "Agent",
@@ -2201,13 +2203,14 @@
       "conversation": "Conversation",
       "knowledge-base": "Knowledge",
       "recent": "Recent",
-      "session": "Work",
+      "session": "Session",
       "topic": "Topic"
     },
     "keyboard": {
       "select": "Select"
     },
     "messageSearch": {
+      "entry": "Messages",
       "hint": "Type to search message content",
       "matchModeLabel": "Match mode",
       "matchModes": {
